Update boarding missions.txt
As a means of making it a little more viable to play small fighters and such early on in the game, I went ahead and came up with a slightly more sophisticated set of boarding and assistance missions that you can get, with a focus on early-game play. A secondary goal was simply to make this part of the game a little more rich and interesting at any point in the game. These changes were only minimally tested, so if you see something that doesn't seem to work, please let me know.

Also, values, script syntax, and dialogue grammar probably need checking. I'm new and doing this in a hurry, so I'm sure I've made. 

Take special note of the mission that is supposed to give you access to a Navy license; I'm not sure whether that is going to work and would appreciate some input.