Condensing "Food" into more useful categories
The former list of "food" categories is a little unusual in how specific it is. It's honestly a little silly to get a mission to transport 80 tons of frozen pizzas or 130 tons of brownie mix. Therefore, I have removed many of the sillier items and instead used broader categories. There are a couple of reasons for this:

 - It removes the silliness of transporting enormous quantities of oddly specific and niche items
 - It is more realistic in that a shipment would include several items (e.g. a truck carrying food in the real world would likely have more than one type of item to a store that requests several items from a supplier).
 - It gets us away from Earth-based food items and leaves room for the player to imagine other types of things that are discovered on other planets. For instance:
     - "Seafood" might include catfish from Earth, or treel from the Babylon 5 universe.
     - "Lifestock" might include cattle from Earth, or "nerfs" from the Star Wars universe.
     - "Alcoholic beverages" might include whiskey from Earth, or Romulan ale from the Star Trek universe.

I believe that there are still plenty of categories to give sufficient variety in missions. If this pull request is well-received and implemented, I am happy to create pull-requests like this one for the other categories (shipping 90 tons of bedpans? Really?).