a few more ships, jobs
bounty hunting jobs now have combat rating unlocks through 6000, with a
stagger on the random range (so that the user doesn’t have an
all-or-nothing type job board) and distributed variants throughout.

Marauder Fury is intended as a persons ship - i’m for "Marauder Max K”,
but i’m sure he’d be just as happy to see Cap’n Pester flying it.

just waiting on the Marauder Bounder sprites for gun/engine port
mapping, and this should be ready to merge…