Fix segfault on problematic paths
Previously, the return value of SDL_Get{Base,Pref}Path was not checked
for validity which could result in segmentation faults if one of several
oddities occurred (e.g. creation of the saves directory failed for lack
of writing permissions on ~/.config on Linux). Now we exit somewhat more
cleanly and provide a useful error message in this scenario.

Fixes #1283.