Give feedback to user about blocked mission
http://steamcommunity.com/app/404410/discussions/0/360671247404843393/?tscn=1470191154

The mission would not complete if the player does not have 40 credits, and the player would also not be informed as to why the mission would not complete. This informs the player that the mission can not be completed at that moment by prompting such with a blocked dialog as well as creating a reminder mission that will make sure that the player still knows to go to Alexandria.

Pretty sure I got this right. Just wrote it up and didn't test it. I'll do that in a bit.