Added message reminder of mission stops; simple implementation of #1383.
This is the "dumb version" talked about in that issue, to get feedback. In
`Engine::EnterSystem()`, a check is made for missions with either
destinations or stopovers in the newly-entered system. If any are found,
`Messages` spits out "You have a stop to make on $PLANET." There's also a
check made on the `Mission`'s `NPCs`, to avoid a reminder for an unfinished
bounty.

Making this more elaborate looks like it would involve duplicating the logic
inside `Mission::CanComplete()` (or ginning up a set of `PlayerInfo` objects
with the planet changed).