OutfitInfoDisplay: Move shots per second to the end of the info block, next to anti-missile damage per shot
Shots/sec isn't relevant to the per second rates for damage, energy,
and heat:  It belongs more with the per-shot data.

For ammo-using weapons, this puts it next to missile strength, which is
relevant to overcoming anti-missile systems.  For AMS, it puts anti-missile
per shot next to rate of fire.  (Still not grouped with range, though.)

I considered adding an anti-missile dmg / sec, but the discrete effects
matter a lot depending on what you're shooting down, I think.
(Overkill vs. just barely needing another shot).

"anti-missile damage / shot:" makes the anti-missile mechanics clearer.
Maybe not needed once the shots / sec is right next to it, but previously
the only clue that it was per-shot was context (being in the per-shot block
of stats).  But it *didn't* have "/ shot" at the end while the other lines
nearby did, so I assumed it was something else.  (I just thought the Hai
turret was way better than Human, like their reactors and shields are.)