Adds a new and bigger Hai cooling system.
This adds a bigger cooling system to be used mainly by Unfettered Hai in their Solifuges. It can replace 3 Hai Williwaw Cooling devices and in fact it has the same cooling power than 3 Williwaws but saves a little bit of space.
The specs are similar to the Liquid Nitrogen Cooler.