Patch for #1826 (Weapon range problem)
Added support for random lifetime and random velocity, so that these now
affect the range in the Outfitter Panel and in AI firing decisions.

Added support for submunition velocities in range calculations.

Added support for a "optimal range" weapon attribute which allows you to
override the range for the purposes of AI firing decisions only.