Additional Mission Options in Free Worlds Start
Hello, all,
I'm a first-time coder here, and I've written up a few additional missions to supplement the Free Worlds start chapter. 

The idea with these missions is twofold:
1. To allow a player who has built up his ship/fleet and has a very high combat rating to test him or herself in a very challenging battle.
2. To provide the player with a more reasonable and defensible course of action in attacking Greenrock. 

How it's supposed to work:
If a player with a combat rating above 5000 completes "fw Pirates 3.2A" aka "Attack Greenrock," a priority mission offer comes up pre-empting the standard "Report to Tomek" mission. If the player declines it and follows Tomek's drunken orders, then the mission string continues as before and (I think) the "Report to Tomek" mission should come up. 

If the player accepts, then he immediately launches to find a pirate armada fleet: four times as many pirate ships as in the previous battle, plus 20 assorted Marauder fleets and a few special ships, including an Alpha carrier. To make matters worse, the rest of the FW attack fleet is gone and he/she must either defeat the fleet(3.2C) or run away. This mission is on a 1-day timer, so running away or landing will fail the mission. Failing this massive combat mission means the player reports to Tomek and the mission string continues as in the previous version.

If the player defeats the pirate armada, he/she lands in covert mode. The story text describes the player freeing the slaves and utterly destroying the pirate's support base. This triggers an event which eliminates all port services from Greenrock. It also starts the new FW Pirates 3.1D, aka "Get the Freed Slaves to Safety." This leads to Longjump, where the player begins the new FW Pirates 4.2 aka "Senate Hearing." 
The character provides some moral arguments in this mission. Although many people, including many senators, cheer the outcome of the player's actions, legal issues force the senate to publicly admonish the player while clandestinely giving him/her a modest pay raise. 

Anyway, please have a look and let me know what you think. Apologies in advance for coding or procedural mistakes; this is my first time playing around with this.