Dangling Modifier Fix (and consistency)
https://owl.english.purdue.edu/owl/resource/597/01/
previously, "as part of the program" was modifying "transport", which is incorrect; there is no clear subject to the introductory phrase "as part of the program". It now has a subject (the volunteer settles).

Also, look at Coalition: Transfer 1 just directly below the modified missions; they are done correctly by default.