Grammar, Formatting and Diction fixes
Addresses several plzrewrite requests, some missing closing backticks. Formatting changes to make it in line with the rest of the game's text (whitespace and such). Changes a couple words here and there for style + effectiveness, also adds another label to address Jafdy's option. Only occurs when you ask about the location of the attack, because it goes into the specifics.