Added gov FW and escort tags
Change for FWC Diplomacy 1C due to lore: (Big Brother) takes off immediately = +waiting.
Possible change for FWC Diplomacy 1C: add accompany save to this NPC, but this would need the `on visit` stuff to be added. Thus not done.

Not changed: FWC Checkmate 1, reason: not sure about the intention of the first two NPCs (Line 653 vs. Line 675) and the very last NPC (Line 819) and not playing this right now. They seem to miss the `escort` personality and the first one the government Free Worlds tag.

Decided for gov Escort in FWC Checkmate Escorts (Line 1040) because their intention is different from 1-mission-escorts. Explicit definition makes clear that I didnt forget it but decided by purpose. Reason is to allow the user to differentiate between regular short term F.W. escorts and the ones that really should survive.

Throughout this file the `escort` tag is often not needed in terms of `npc accompany` but it gives the user a better feeling for his strength if the fleets which he leads into battle are shown in his escort list.