Added escort tags
I added the escort tags only in the first mission, but not in "FW Pirates 4", Line 1422+
Reason: The first mission is described with
`description "Travel with the Free Worlds militia to the <system> system and drive off the Navy fleet that is occupying the system."`
i.e. the user is fully aware of the fleet and travels along with it.
In "FW Pirates 4" the user is told about some reinforcements but may be should not be fully aware of their numbers?