Clarify ship attribute table meaning
Addresses endless-sky/endless-sky#2406 by changing the shipInfo column label from "heat" to "heat gen".
The tooltip does indicate that this column refers to heat generation rate, but it is very easy to misinterpret the column's meaning when it is labeled "heat" and its neighbor is labeled "energy", since the energy value in the "max" row is actually the maximum energy, while the "heat" value in the "max" row is a much more complex value.