Added condition "weapon capacity" where "net worth" is calculated
While "cargo space" and "passenger space" reflect civilian capacities of the players fleet, the "weapon capacity" reflects it's maximum total armament.
This can be useful for mission conditions. Particularly it is now possible to compare between "cargo space" and "weapon capacity" as a rough estimation whether the player has a rather military fleet, mostly of warships, or a rather peaceful fleet, mostly of freighters.