Jump commands cancel in-progress MOVE_TO orders
[x] Clear MOVE_TO order if it was issued before jumping (fixes original issue #2382)

If you jump (or are executing a travel plan), any ships which are still moving to the position specified by a MOVE_TO order will abort the command and follow you to your new system. Any ships which completed the order will remain in the current system at the specified location.