Carried fighters/drones no longer clear orders needing a target.
This commit fixes an issue where issued fleet orders that require a target (such as GATHER or FIGHT/FINISH_OFF) would not "toggle" if the selected group of escorts included carried fighters.
This has the added benefit of being able to set a combat target prior to deploying fighters/drones and not needing to re-issue that target after deploying fighters/drones.

Closes #2608