Implemented outfit and ship tags
Any outfit or ship can have a series of tags, e.g. tags alien remnant
warship, and each tag is turned into auto-conditions that can be used
for mission logic. This can allow, for instance, a mission that triggers
if you have Quarg technology, a sponsorship deal if you use Lovelace
weapons, or a mission requiring you have the ability to scan cargo
(instead of specifically the Cargo Scanner outfit).