Fleet list: hitting 'up' with no selection now goes to the end of the fleet
At present, if you open Player Info and hit 'up', the bottom ship is
selected but the page will scroll only one page down from the top. With
this tweak, it will now scroll down as many pages as needed.

Note that as before, between the first ship (index 0) and the last is
'index -1', which is no ships selected. The current view won't move when
you move to index -1.