Ships can have a custom "swizzle" property (#2629)
* Ships can have a custom "swizzle" property

The nomenclature is the word "swizzle" (quotes unnecessary) followed by
a number between 0 and 8, at the same level of indentation as sprite or
attributes. Use 0 for no swizzling of the original sprite.

* Defined key in Ship::Load

* Custom swizzle corrections

Invalid ship swizzle values will be ignored. Also fixed a missing
if(government) check in Ship::Place, introduced in previous commit.