Right-clicking the Radar display will issue a scaled MOVE_TO order
Current behavior is to allow clicking "behind" or through the radar display, e.g. a right-click acts as though the radar display is not present.

With this PR, right-clicking the radar display will send a MOVE_TO order to your selected escorts, with the scaled target position.
This allows ordering your escorts up to 128 / 0.025 = 5120 units away from your flagship's current position in the system, without requiring you to zoom out.