Heroic personality no longer targets any hostile in system
Heroic ships will instead target any hostile within 5000 of themselves, or an ally's target that is within 5000 of it
Unconstrained ships will target any hostile, so long as it is not too powerful.
Heroic unconstrained ships will target any hostile no matter how strong, and no matter where it is in the system.
Ref #2704