Restore follow-parent behavior for hostile NPCs which began patrolling
In 4ab74b5bc8 hostile, non-staying NPCs which could not target any ships in a system were instructed to patrol it instead of jump away and then jump back. However, if the player left the system while remaining untargetable, the NPCs would get stuck landing if they were targeting a non-planet at the time of your jump.
This commit resolves the logic bug in MoveEscort that assumed the ship's stellar target was a landable planet by verifying that it is indeed a landable planet. If it is not, the ship will clear the target and resume tracking down the player.