Lore friendly Quarg Rep
During the Wanderers campaign, the Quarg mention their dislike for the Pug, and the Pug express similar feelings. When talking to the Coalition/Heliarchs, they mention their dislike for the Quarg, and you can assume that the feeling is mutual. With this in mind, I changed the Quarg/Pug and Quarg/Coalition/Heliarch so that killing Pug, gives a small rep increase to Quarg, and visa versa, and killing Quarg increases Coalition/Heliarch rep and visa versa.