Consider the next refuel system when an escort is traveling to the parent.
An escort was only checking if it had enough fuel to reach the next
system, so it could get stuck when it had to travel through two
consecutive systems where it couldn't refuel.

Now it checks how much fuel is required to reach the next refuel system
on the travel route to decide if it should refuel in the current system.