Add flight check information of your ships to ShipInfoDisplay.
The "flight check:" line comes after the attributes table (energy and
heat). It will have a value of "complete" if everything is fine and the
number of warnings if problems were detected. Each problem will show as a
paragraph in the tooltip.

The extra line also fits in the ShipInfoPanel.