Always show flight check line.
The flight check line will also show in the shipyard and map's shipyard.

Quoting tehhowch, "If a stock ship being sold is not ready to fly, that
should be communicated to the player. This is beneficial for
plugin-makers, as a content sanity check, or as a gameplay mechanic
(e.g. selling discount ships which are missing parts)."