Use selected item as reference in ShopPanel.
In the outfitter, when you have an outfit selected and choose a different
ship it will keep the outfit in the same screen height.
This is needed because when a new ship is selected it can add or remove
outfits before the selected outfit, causing the selected outfit to be
placed in a different position.

It is also able to auto-scroll down and tries to make the image visible.