Update ship warnings.
Warnings can be serious or not.
In the list with your ships of the shop, the warnings are packed together
into a single icons and tooltip.
In your currently selected ship in the shop, the warning icons and
tooltips are displayed in a line at the bottom of the ship icon.
In ShipInfoPanel the warning icons and tooltips are display in a line at
the bottom of the ship image.