Rework flight check conversation.
It handles 1 warning, even if unknown.
It uses player conditions as temporary input variables.
It allows 4 actions, go to the outfitter, abort the take off,
park the ship, and dismiss the warning.

Abort the take off is always available.
Park a ship is available if it's not the flagship.
Dismiss the warning is available if it's not a serious warning.
Go to the outfitter is available if you can access the service
and the ship is in the current system.