Show depreciated stock of items that can be bought.
An outfitter has a unlimited stock of outfits that can be bought. When you
bought and sold back an outfit it would be placed with the depreciated
stock which would give incorrect stock numbers (if displayed).

Now new outfits are sent back to the infinite stock and the depreciated
stock is shown.