Keep partially collected flotsam (#2955)
* Keep partially collected flotsam

The leftover units would vanish when only part of the flotsam fits the
cargo hold. This means that you can lose cargo just by transfering it
to another ship of your fleet (jettison cargo, then guide other ships to
the flotsam).

This commit changes that behaviour, only removing the flotsam when there
are no more units. Only 1 ship can collect from the flotsam on each step.

* Jettison partially collected flotsam from the collector ship

Place the flotsam as if it was collected, partially stripped,
and then thrown back out.

* Replace placement of partial flotsam with an adjustment in velocity.

Flotsam transfer logic is moved to Flotsam::TransferTo and the flotsam
velocity is stabilized in proportion to the amount being transferred.

* Cast to double with static_cast.