all hails.txt the best /data/ of all
This might be an on going project, as it's fun to make these. At least until they all have quite a few options. If I need to work on this somewhere else to do the same thing, then I shall. Or I'll just upload less often and check over it more.