Same as my hails.txt but organized/searchable
This is the same file that I submitted on the other commit, but rather than being random, I sorted it. I didn't commit to the previous one in case you guys for some reason don't want it to be sorted and searchable.

#navyfriendly
#navyhostile
#freeworldsfriendly
#freeworldshostile
#other
are a few examples. Will make it easier to ensure there's no duplicates now.