Adding inhibitor.wav
Adding an audio file for the Inhibitor Cannon, to replace its use of particle.wav

With the rapid rate of fire and its typical use in large numbers, the particle sound is jarring on the Inhibitor Cannon.

The intent of this replacement sound is not only to give it a unique audio, but also one that is less obtrusive when used rapid fire and in large numbers.

This audio is derived from the Korath Detainer (the supposed source of Remnant Inhibitor technology), but sped up and pitch altered to fit a smaller weapon.