Add assassination missions to Pirate Worlds
It would seem logical that pirates would have some enemies that they would need taken care of. A few "problems" that need "solving". I have made 4 assassination jobs (2 political, 2 military), which vary in difficulty, and the target's faction changes based on where the job is accepted.

Free Worlds targets in the south
Syndicate targets in the core
Republic targets in the north

NOTE: I realize that the northern targets may spawn in the pirate region. I don't know of any way to prevent this from happening