Make settings available directly as fields
Previously, settings were being accessed by calling a method and
passing a name (i.e. a string literal), which means that every time a
setting was accessed, its name needed to be looked up in a string
map. This is an unnecessary waste of cycles, and becomes prohibitive
to introducing preferences that are to be accessed many times per
frame.

Improve on this by exposing the preferences directly as fields of a
static class, and update the rest of the code to use them. Making them
available in this manner removes the string map lookup, method call
overhead, as well as any address calculations (due to them being
static). Aside from the performance improvements, this has the benefit
that a typo in a preference name will result in a compilation error,
instead of silently retrieving the default value (false).

* Preferences.h:

  - Add public fields for settings whose value can be used as-is

  - Replace Has/Set methods with a single Get method, returning a
    writable bool reference

  - Replace static methods with non-static ones, instead declaring a
    class instance (defined in the .cpp file). This is so all
    variables can be reinitialized at once to their initial values
    before loading settings; an alternative would be to move the
    settings to a subtype, at the cost of making access to the
    variables more cumbersome (e.g. Preferences::values.showMinimap).

* Preferences.cpp:

  - Add/update implementations for the above methods

  - Replace the map<string,bool> with statically-initialized
    map<string,bool*> for name lookup

  - Move setting names into an array, so that they are queryable by
    e.g. Name()

* PreferencesPanel.cpp:

  - Update to use the new Preferences API

  - Add some missing "else" keywords

* Help.cpp, Help.h:

  - Introduce Help::FindTopic needed by the Preferences class to load
    the state of help texts.

* AI.cpp, DrawList.cpp, Engine.cpp, MainPanel.cpp, MapPanel.cpp,
  PlayerInfo.cpp, Sprite.cpp, main.cpp:

  - Use the Preferences instance fields to set and query settings
    instead of calling Has/Set with string literals