Add electromagnetic storm missions
The Ember Waste is a unique region of space, but it's mostly just a different colour and has lots of special wormholes. What if there were missions that took place here, that involve electromagnetic storms? Essentially, the concept is that overtime, charged particles in the waste reach higher concentrations in certain regions, resulting in electromagnetic storms. It adds a new twist to deal with in a combat mission that I've included in the Ember Waste. This is probably far from ideal in it's current state, but it's a proof of concept.