Fixing tab-to-spaces errors
It appears that a bunch of the stuff related to adding in the Korath Hunter (basically all the copied code) converted the tabs into spaces. This patch fixes that so they are all back to being tabs. Also deleted a bunch of extraneous tab/spaces trailing the content on some lines.