Prevent messing up the FW storyline (#3749)
* Made Republic hostile to Korath exiles again

It is the duty of the Republic Navy to protect the innocent people; the Korath exiles are attacking the latter. It really doesn't make sense for the Republic to ignore the exiles. And for messing the storyline, other commits will prevent that.

* Prevent messing up the storyline

Made the Republic super duper very very much a lot difficult to be friendly to the player by making them more hostile that it will take a lot of time to kill pirates to make them friendly again

* Prevent messing up the storyline v2

Adjusted the attitude of the Republic to all factions to (-) .1, instead of lowering the reputation to ridiculous amounts

* Prevent messing up the storyline v2.1

(-).01 instead of (-).1

* Republic no longer hostile to Korath exiles

Per https://github.com/endless-sky/endless-sky/pull/3749#issuecomment-393786167

The Korath thing will be separated to another branch instead (and another PR)

* Removed anything Korath

The Republic is no longer hostile to Korath exiles