Removing Korath Hunter
The creator of the Remnant Hunter can't be contacted for unknown reasons, so I can't get any additional files (or copyright info) for it.

Removing it for now (replacing with higher numbers of Chasers), but with the intention of adding the Hunter back in once the creator is availble and can provide the rest of the info & files.

Note: If someone else introduces a Korath Interdictor or Light Warship, I will use in the same way I was using the Hunter.