New mission involving the intro banker
This is the first time I've ever done anything remotely close to writing or coding, so I double checked everything but there's a very realistic chance I made an error, which I would be happy to fix. The mission involves fleshing out the loan broker who gives you the loan to purchase your first ship and transporting some of his hired thugs to a Syndicate planet to deal with a challenging client. Hope it looks good.