Minor Mission Wording Changes
Added some minor wording changes for better connotation and implicational consistency.

First change was to change it from an order to a teaching moment. Considering the way James acts, I think that this a bit more like what would happen.
Second change is just for a little bit of chronological consistency. You can't really be taking payment from somebody if they're currently being pulled away and talked to. If you say after it puts it together a bit more.
Third change is for more connotation consistency. Walk is very casual while march is very uptight. Pass is more neutral and doesn't contradict like walk and march does.
Fourth change should be on, not of.