Dialogue improvement
The navy has a diverse selection of people with different religions: Christian, Islam and Buddhism are referenced in the game. Requiring a Buddhist or Muslim to swear on the Christian god is counter productive to diplomacy. Christianity declares Oath Swearing taboo as Jesus spoke against the practice so including God is taboo as well. The only commonality is their sense of honor which they consider sacred.