Update remnant missions.txt
Added a "Remnant: Learn Sign 3: done" check on Salvage 5. Basically, once the player has completed the first two missions of learning sign they can continue to doing the Salvage missions, but they won't be able to progress past Salvage 4 (and the unlocked lurker bounty) until they've completed the learning sign timer.

This is intended to create a sort of split, where there are a variety of missions that can be done after having simply acquired the license; then a handful more that can be done after having learned the basics of the language, and more that will be unlocked after full conversational fluency has been reached.