Misc changes
Move some events to trigger when they are actually mentioned in the story

Salvage 3 now mentions to go to the spaceport for Salvage 4

Changed some dialog in Salvage 1 to better fit the description of the technology your brought to the Remnant no matter what it was.

Cut down on two parts that were far too novel-y. (To paraphrase MZ, this is a game, not a novel. Quote, "In general, I try to make mission text as concise as possible and make sure that every sentence the player is forced to read is worthwhile.")