Increase crew attack and defense for Coalition species
The Coalition is made up of species with increased affinity for close-quarters combat. They all move fast. Arachnids and beetles (animals the Arach and the Kimek are based upon) are carnivorous by design and have the benefit of their sturdy carapaces (natural armour) and natural weapons. Even Saryds would benefit from thick hides and strong kicks.

This proposed change reflects that.

Coalition civilians aren't taught to fight well, so they would have to rely on their natural instincts, There's also a possibility that a ship might be crewed by an all-Saryd crew (and they wouldn't be the ideal troops for fighting inside ships). Heliarch crews, however, are always mixed and are trained well to work in concert for their advantage.