Modify human warship H2H to better fit their roles
Changed the starting point to "1/4 required crew, rounded down,
minimum 1". From there, ships that are described as designed for defense
of planets or fleets get fewer rifles, since they're not expected to
board or be boarded. Ships designed for adventuring/aggression get more.
Ships largely for cargo or likely to get boarded keep the default.

This also includes slightly raising the FW military ships' H2H, but not
as high as the Republic numbers. The already high Republic numbers
remain unmodified.