Fixing lore issues
Fixed lore issues related to the Shield and timeline of Korath conflicts. The former by adding to the description that the gascraft uses an antique hard-shell system that must be manually raised/lowered; and the latter by unspecifying the nature of the ground-based threat. All that is specified is that it made having a area of effect weapon that could penetrate through walls was very useful during a period when building-to-building clearing operations were unfortunately common.