Indigenous -> Void Sprite
This commit breaks off the Void Sprites from being generic animals to being a separate government. The Indigenous file is removed and the contents thereof are now in "void sprite ships.txt", so that when future space wildlife is added, those can use the indigenous file.

Note: The void sprites will continue to appear as "Indigenous" until the player reaches the appropriate mission, after which the void sprites will cease to appear as indigenous, and will be labeled as "Void Sprite"

Changes all void sprites to be "automata" so they can fly without crew, removes bunks & crew, and changes cargo space so that none of the Void Sprites could ever be equipped with a control center. All this is basically a long-winded way of saying "At some point I want to allow them to be purchasable/hireable as ships in your fleet, but as sentient beings that weren't built to be ships for people of any kind, they should never, ever be able to be directly flown by the player."