fw epilogue edit
When testing the Paradise Fortune mission chain, FW Epilogue: Alondo offered, leading to a situation where I sat down and had a beer with Alondo while the Republic was hot on my heels.

The mission Paradise Fortune 4 has a branch that I assume was intended to handle this, but it does nothing to stop the mission from being offered, only changes the beginning dialog.

This PR handles this oddity by removing the ineffective branch from Paradise Fortune 4, and marks the all the epilogue missions minor to avoid the need for any special dialog for this case or any other.