Changed the contribution notice in the credits
Moved it to the top so that it's easier to notice.
Added reference to reporting bugs and giving feedback.
Added reference to the Steam forums and Discord server, as those are also places where we take feedback and bug reports (which we funnel to GitHub regardless if we find an actual bug or get good feedback).
Ref: #4295