Add squadrons (groups of similar ships)
Add support for squadrons, groups of similar ships.

Initial usage of the squadron would be to define larger NPC fleets where you
want multiple identical ships in a single fleet.
Another usage would be to have missions where the player is always escorted by
a certain ship or group of ships, and when such a ship gets destroyed, then
new instances of the ship will appear once the player jumps (starting from
a planet where the ships hull is for sale).

Later usage would be to also allow the player to define and use squadrons.
The player can command each ship in a squadron he/she owns just like any other
ship, but the first ship in the squadron acts as squadron-leader. So any other
squadron ship will move to the squadron leaders location if the other ship has
no other commands.

A squadron could also consist of 0 ships if all ships in the squadron are
destroyed. The squadron definition then still exists and the player can land
on a planet where the squadron ships can be bought to get new ships in the
squadron from the shipyard in the same way as that ammunition can be bought
(buy ships to re-equip your squadrons? Question at the shipyard.).

Buying squadron ships doesn't need to be an exact match with ships in the
shipyard. If the hulls and relevant outfits are available (either in the
outfitter or as cargo in the players cargo-hold), then the squadron ships
can automatically be bought and built for the player.

Squadrons owned by the player can only be started from a ship that the player
owns and that is on the same planet as where the player has landed.