Updating the gascraft
Added built-in thrust & turn equivalent to a set of crucible engines

Removed the set of crucible engines & equilvant amount of space.

This change means that the player cannot swap these components out for the Coalition's small thrust module and small steering module, thus ensuring the player can't use those to outfit the gascraft with a cloaking device.