Updating Heavy Freighter
This update corrects the Mule:
- Changed classification to Heavy Freighter
- Changed description to indicate that the Republic officials were tempted to classify it as a warship instead of actually *being* classified as a warship.

These changes bring the mule in line with the Behemoth (which has more hull, more shields, and more weapon space than the mule) and the Bulk Freighter (which also has significantly more hull and somewhat less shields (but still more combined health) and as much weapon capacity as some medium warships and more than most.). To a lesser extent, this also brings it in line with the Argosy, which has more hull, shields, and weapon space than many light warships.

One of the reasons for this change is to increase the number of options for players who wish to play as merchants who happen to be good at combat. It is hard to pretend to be a merchant when one is flying a heavy warship, no matter the justifications. Thus this change reclassifies the mule as a heavy freighter, and modifies the description to give it some ambiguity: It is more interesting to look at the mule and see how regulators might have classified it as a warship; rather than to look at it and see a subpar warship that only barely made it into the category.

Also, while what the regulators term a warship may have been relevant back when licenses for warships were a thing, now that they are not, keeping the mule as a warship simply muddies the waters and hides a ship that merchant players may wish to fly in a section where they might not look when comparing freighter capacity.