Add option for built-in guns and turrets
Hyperdrives, hull-generation, cloaking and many other options can already
be built-in into ship hulls. This change adds an option for guns and
turrets to be built-in into the hull of the spaceship it is installed in.

Built-in weapons cannot be removed from the ship they are installed in.
And selling the ship that has built-in weapons does not put the built-in
weapon in stock.

Signed-off-by: Peter van der Meer <peter.vd.meer@gmail.com>