Change bay type identification to categories
Change the ways bay types are identified from boolean to the category of
the ship that is to be carried in the bay.

This will make it easier to add new bay types in the future and/or to
load bay types from data files instead of having them hardcoded.

Signed-off-by: Peter van der Meer <peter.vd.meer@gmail.com>