Fixed reloading issue and fixed hardpoint load issue
Built-in weapons don't show up in outfits, so changed the code to handle
weapon reloads through hardpoints instead of through outfits.

Also fixed a load issue for partially specified turrets (in case of
ship variants).

And moved calculation of equipped weapons to FinishLoading function of
Ship. It is only used there, so no need to keep it available after
loading.

Signed-off-by: Peter van der Meer <peter.vd.meer@gmail.com>